+++ pkgs/package_config/lib/src/package_config.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2019, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2019, the Dart project authors. Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
@@ -6,6 +6,7 @@ import 'dart:typed_data';
 
 import 'package:meta/meta.dart' show sealed;
 
+import 'constants.dart' as constants;
 import 'errors.dart';
 import 'package_config_json.dart';
 import 'util.dart';
@@ -20,10 +21,10 @@ import 'util.dart';
 @sealed
 abstract class PackageConfig {
   /// The lowest configuration version currently supported.
-  static const int minVersion = 2;
+  static const int minVersion = constants.minConfigVersion;
 
   /// The highest configuration version currently supported.
-  static const int maxVersion = 2;
+  static const int maxVersion = constants.maxConfigVersion;
 
   /// An empty package configuration.
   ///