+++ pkgs/package_config/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 2.3.0-wip
 
+- Adds discovery API to find both a configuration and its location:
+  `findPackageConfigAndFile` and `findPackageConfigAndUri`.
+
 - Removes support for the `.packages` file.
   The Dart SDK no longer supports that file, and no new `.packages` files
   will be generated.
@@ -13,7 +16,7 @@
   - Old functions still exists as deprecated, forwarding to the new
     functions without the `preferNewest` argument.
 
-  Also makes `PackageConfig`, `Package` and `LanguageVersion` `@sealed` classes,
+  Makes `PackageConfig`, `Package` and `LanguageVersion` `@sealed` classes,
   in preparation for making them `final` in a future update.
 
 - Adds `PackageConfig.minVersion` to complement `.maxVersion`.